add rebuild flag to compose

+++ docker/docker_setup.sh
@@ -7,9 +7,17 @@
 # License: MIT
 # ================================================================
 
-docker stop $(docker ps -q)
-docker rm -f $(docker ps -aq)
+for arg in "$@"; do
+  if [[ "$arg" == "--build" || "$arg" == "-b" ]]; then
+    BUILD_FLAG="--build"
+  fi
+done
 
-docker system prune -af
-
-docker compose -f docker-compose.yml up --build -d
+if [[ -n "$BUILD_FLAG" ]]; then
+  docker stop $(docker ps -q)
+  docker rm -f $(docker ps -aq)
+  docker system prune -af
+  docker compose -f docker-compose.yml up --build -d
+else
+  docker compose -f docker-compose.yml up -d
+fi
